data/requirements.txt
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
requests==2.32.5
rsa==4.9.1
uritemplate==4.2.0
urllib3==2.5.0
